requirements.txt
fastapi>=0.115.2
uvicorn[standard]>=0.27.1
python-multipart>=0.0.18
openai>=1.30.0
python-dotenv==1.0.1
pydantic>=2.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0
torch>=2.0.0
scikit-learn>=1.3.0
gradio>=4.0.0